Changed netgen/setup patches, colab spice, and CVC power filename
netgen setup patches:
- delete permute terminals for npn, pnp
- delete parallel reduction for npn, pnp
- change cap_rfcmim to rfcim
Since spice files are now generated from xschem, inline spice file cell deleted
CVC execution is based on the existence of a power file suffixed with top source name. The top source name was changed to the project name, but the CVC power file was suffixed with the top layout name. This commit corrects that for 160GHz and 40GHz designs.

Signed-off-by: Mitch Bailey <d.mitch.bailey@gmail.com>